feat: styles for composer and success
