--- src/components/homepage/Composer.tsx
@@ -1,746 +1,0 @@
-import React from "react";
-
-export default function Composer() {
-  return (
-    <div className="py-24 px-12 grid grid-cols-2">
-      <div className="flex items-center justify-center">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="500"
-          height="500"
-          fill="none"
-          viewBox="0 0 630 630"
-        >
-          <g clipPath="url(#clip0_11024_3769)">
-            <rect
-              width="134.4"
-              height="134.4"
-              x="483"
-              y="248.85"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="12.6"
-              y="248.85"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="413.7"
-              y="82.95"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="81.9"
-              y="414.75"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="247.8"
-              y="13.65"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="247.8"
-              y="484.05"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="81.9"
-              y="82.95"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <rect
-              width="134.4"
-              height="134.4"
-              x="413.7"
-              y="414.75"
-              fill="#D9D9D9"
-              rx="64"
-            />
-            <path fill="#DEE4F3" d="M79.8 310.8h470.4v8.4H79.8z" />
-            <path
-              fill="#DEE4F3"
-              d="m145.719 478.342 332.623-332.623 5.94 5.94-332.623 332.623z"
-            />
-            <path fill="#DEE4F3" d="M310.8 550.2V79.8h8.4v470.4z" />
-            <path
-              fill="#DEE4F3"
-              d="M478.342 484.281 145.719 151.658l5.94-5.94 332.623 332.623z"
-            />
-            <g filter="url(#filter0_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="81.9"
-                y="82.95"
-                fill="#EF6351"
-                rx="67.2"
-              />
-              <path
-                fill="#fff"
-                d="M158.33 141.009a2.106 2.106 0 0 0 1.472-.587 2.07 2.07 0 0 0 .634-1.439 2.097 2.097 0 0 0-.623-1.468 2.144 2.144 0 0 0-1.483-.617 2.111 2.111 0 0 0-1.453.628c-.382.39-.595.913-.592 1.457.004.536.221 1.049.604 1.427.382.38.9.595 1.441.599ZM172.64 142.957h5.079v.399a2.75 2.75 0 0 0-.806.276c-.464.215-1.393 1.258-2.105 3.005-1.208 3.036-2.476 6.622-3.808 10.731l-.496 1.559c-.217.704-.403 1.164-.495 1.441-.093.278-.218.705-.433 1.225a6.863 6.863 0 0 1-.472 1.135c-.249.46-.713 1.402-1.085 1.686-.588.491-1.456 1.043-2.538.951-2.106 0-3.685-1.559-3.716-3.402 0-1.258.805-2.086 2.013-2.086.867 0 1.64.468 1.64 1.44 0 .707-.712 1.442-.712 1.81 0 .95.557 1.402 1.642 1.402.865 0 1.574-.551 2.104-1.655.713-1.257.774-2.636.186-4.17l-4.457-11.681c-1.023-2.697-1.798-3.584-2.756-3.679v-.399h7.027v.399c-.836.092-1.268.583-1.268 1.471 0 .644.247 1.625.711 2.883l.836 2.268c.96 2.453 1.731 4.599 2.137 5.948.897-2.789 1.731-5.242 2.539-7.388.588-1.595.867-2.729.867-3.404 0-1.195-.65-1.747-1.642-1.747l.008-.418ZM123.749 157.889c-.897-.156-1.672-.951-1.672-2.637v-20.327l-5.108 1.103v.43c.867-.092 1.731.674 1.731 2.299v16.495c0 1.594-.867 2.514-1.731 2.637a2.733 2.733 0 0 1-1.536-.197c-.687-.296-1.26-.824-1.744-1.613l-2.361-3.74a4.34 4.34 0 0 0-2.11-1.732 4.387 4.387 0 0 0-2.738-.138l2.663-2.913c2.012-2.207 3.87-3.617 5.635-4.2v-.399h-5.858v.399c1.517.583 1.424 1.87-.315 3.896l-3.746 4.322v-16.649l-5.109 1.103v.43c.867 0 1.732.857 1.732 2.36v16.434c0 1.809-.836 2.514-1.732 2.637v.399h6.787v-.399c-1.114-.156-1.672-1.012-1.672-2.637v-3.034l1.455-1.595 3.522 5.732c.836 1.381 1.61 1.933 2.849 1.933h11.788v-.312s-.336-.023-.73-.087ZM137.813 155.95v-6.857c-.068-4.484-1.966-6.525-6.313-6.525a6.215 6.215 0 0 0-3.838 1.287c-1.146.858-1.703 1.84-1.703 2.975 0 1.103.618 1.931 1.64 1.931 1.084 0 1.859-.613 1.859-1.471 0-.644-.435-1.534-.435-2.146 0-1.105.838-2.056 2.292-2.056 1.859 0 3.188 1.381 3.188 4.446v1.839l-1.547.368c-.805.156-1.485.312-2.013.468-.527.156-1.207.367-2.013.673-1.61.614-2.475 1.197-3.22 2.3a3.197 3.197 0 0 0-.557 1.81c0 2.543 1.796 3.709 4.305 3.709 1.981 0 4.092-1.043 5.045-3.005.014.619.164 1.227.44 1.783 1.049 2.104 4.526.86 4.526.86v-.399c-1.522.236-1.648-1.557-1.656-1.99Zm-3.305-2.123c0 1.012-.371 1.839-1.114 2.422-.712.583-1.454.89-2.229.89-1.517 0-2.509-.982-2.509-2.791 0-.857.473-1.655.868-2.115.312-.317.68-.577 1.084-.767.526-.276.787-.417 1.146-.582l1.416-.521c.712-.276 1.145-.429 1.33-.521l.008 3.985ZM202.913 142.957h-12.042v-8.032h12.042l-2.528 4.016 2.528 4.016ZM178.436 156.336a8.013 8.013 0 0 1-2.297-5.711 7.909 7.909 0 0 1 2.297-5.682c1.518-1.596 3.379-2.396 5.551-2.396 2.139 0 4 .8 5.519 2.396a7.81 7.81 0 0 1 2.327 5.682 7.899 7.899 0 0 1-2.327 5.711c-1.519 1.559-3.38 2.365-5.519 2.365-2.172 0-4.033-.798-5.551-2.365Zm8.341-11.138c-.612-1.186-1.416-1.881-2.361-2.068-1.915-.382-3.609 1.568-4.24 4.668a14.582 14.582 0 0 0-.193 4.222 11.624 11.624 0 0 0 1.167 4.052c.614 1.187 1.417 1.881 2.361 2.068 1.916.382 3.658-1.651 4.294-4.782.527-2.616.26-5.803-1.036-8.161l.008.001Z"
-              />
-              <path
-                fill="#fff"
-                d="M160.127 155.252v-12.295h-10.9v.369c1.456.216 2.149 1.303 1.488 3.067-3.406 9.167-3.189 8.756-3.406 9.491-.217-.705-.713-2.44-1.518-4.616-.804-2.175-1.331-3.619-1.548-4.262-.836-2.545-.557-3.496.806-3.648v-.399h-7.061v.399c1.053.215 1.982 1.41 2.756 3.556l1.084 2.79c1.192 3.014 2.591 7.168 3.059 8.585h2.346c.755-2.182 3.787-10.889 4.193-11.774.438-1.008.933-1.772 1.487-2.3a2.79 2.79 0 0 1 2.1-.87s1.701 0 1.701 1.626V155.252c0 1.714-.836 2.514-1.703 2.637v.399h6.751v-.399c-.893-.123-1.635-.92-1.635-2.637Z"
-              />
-            </g>
-            <g filter="url(#filter1_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="483"
-                y="248.85"
-                fill="#0E1521"
-                rx="67.2"
-              />
-              <path
-                fill="#F9AB00"
-                d="M587.979 347.014c0 5.623-4.567 10.369-10.193 10.369h-1.228c-5.274-.704-9.14-5.447-8.964-10.721v-61.157c-.176-5.272 3.69-9.841 8.964-10.719 5.623-.703 10.72 3.339 11.421 8.962v63.266Z"
-              />
-              <path
-                fill="#E37400"
-                d="M523.483 336.997c5.625 0 10.193 4.57 10.193 10.193 0 5.623-4.568 10.193-10.193 10.193-5.623 0-10.193-4.57-10.193-10.193-.176-5.623 4.57-10.193 10.193-10.193Zm26.888-31.105c-5.623.351-10.017 5.096-10.017 10.72v27.591c0 7.382 3.339 11.949 8.085 13.004 5.622 1.056 10.896-2.459 12.124-8.085.176-.7.176-1.404.176-2.107V316.26c0-5.623-4.567-10.368-10.193-10.368h-.175Z"
-              />
-            </g>
-            <g filter="url(#filter2_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="81.9"
-                y="414.75"
-                fill="#212326"
-                rx="67.2"
-              />
-              <path
-                fill="#95BF47"
-                d="M179.124 452.899a.922.922 0 0 0-.841-.774c-.347-.027-7.155-.132-7.155-.132s-5.69-5.465-6.25-6.016c-.564-.559-1.661-.389-2.089-.265l-2.859.879a19.572 19.572 0 0 0-1.37-3.326c-2.025-3.824-4.995-5.846-8.58-5.851h-.013c-.25 0-.496.024-.746.047a9.584 9.584 0 0 0-.324-.372c-1.563-1.654-3.564-2.457-5.963-2.386-4.634.136-9.248 3.438-12.984 9.313-2.632 4.129-4.634 9.321-5.204 13.34l-9.119 2.791c-2.684.834-2.768.916-3.119 3.415-.26 1.89-7.283 55.582-7.283 55.582l58.86 10.056 25.515-6.269s-10.412-69.557-10.476-70.032Zm-22.14-5.407-4.57 1.396c-.034-2.318-.31-5.541-1.404-8.329 3.513.659 5.245 4.591 5.974 6.933Zm-7.648 2.338-9.828 3.011c.948-3.599 2.751-7.176 4.961-9.521.824-.874 1.975-1.848 3.341-2.404 1.28 2.643 1.556 6.387 1.526 8.914Zm-6.311-12.082c1.087-.021 2.004.215 2.784.724-1.249.643-2.46 1.565-3.594 2.769-2.943 3.119-5.198 7.965-6.096 12.632l-8.072 2.47c1.593-7.353 7.829-18.385 14.978-18.595Z"
-              />
-              <path
-                fill="#5E8E3E"
-                d="M178.287 452.127c-.348-.029-7.155-.131-7.155-.131s-5.691-5.463-6.254-6.017a1.397 1.397 0 0 0-.79-.362V529.2l25.512-6.269s-10.409-69.557-10.476-70.029a.93.93 0 0 0-.837-.775Z"
-              />
-              <path
-                fill="#fff"
-                d="m149.947 465.092-2.96 10.96s-3.308-1.486-7.223-1.247c-5.747.362-5.808 3.943-5.751 4.84.314 4.901 13.355 5.975 14.088 17.449.574 9.027-4.847 15.206-12.656 15.695-9.376.582-14.54-4.883-14.54-4.883l1.988-8.353s5.197 3.875 9.355 3.615c2.717-.168 3.686-2.355 3.588-3.898-.405-6.392-11.03-6.014-11.698-16.517-.567-8.838 5.306-17.792 18.262-18.603 4.995-.318 7.547.945 7.547.945v-.003Z"
-              />
-            </g>
-            <g filter="url(#filter3_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="247.8"
-                y="13.65"
-                fill="#FFF9F4"
-                rx="67.2"
-              />
-              <path
-                fill="#FF9090"
-                fillRule="evenodd"
-                d="M345.463 120.75h4.733c1.097 0 2.021-.941 2.079-2.057V66.199c0-5.172-4.098-9.346-9.18-9.346h-6.06c-1.096 0-2.02-.94-2.02-2.058v-4.76c0-5.174-4.099-9.347-9.178-9.347h-39.196c-5.08 0-9.178 4.173-9.178 9.346v52.728c0 5.114 4.098 9.288 9.178 9.347h45.43c.404 0 .808.117 1.155.353l11.082 7.935a2.06 2.06 0 0 0 1.155.353Zm-11.025-16.106h-47.682c-1.154 0-2.02-.882-2.02-2.058V50.152c0-1.117.924-2.058 2.02-2.058h38.965c1.097 0 2.021.94 2.021 2.058v4.643c0 1.117-.924 2.058-2.021 2.058h-29.036c-1.097 0-2.02.94-2.02 2.057v34.976c0 1.117.923 2.058 2.02 2.058h36.252c1.097 0 2.02-.941 2.02-2.058V66.317c0-1.117.924-2.057 2.021-2.057h5.944c1.097 0 2.021.94 2.021 2.057v42.912c0 1.057-1.155 1.645-1.96 1.057l-7.391-5.29a2.053 2.053 0 0 0-1.154-.352Zm-7.735-16.166H302.92a.982.982 0 0 1-.982-1v-22.22c0-.528.404-.998.982-.998h23.783c.519 0 .981.411.981.999v22.22c0 .529-.404.999-.981.999Z"
-                clipRule="evenodd"
-              />
-            </g>
-            <g filter="url(#filter4_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="413.7"
-                y="414.75"
-                fill="#FFD250"
-                rx="67.2"
-              />
-              <path
-                fill="#664F44"
-                d="M519.773 520.175c-.299.299-.672.446-1.118.446-.449 0-.821-.147-1.118-.446a1.518 1.518 0 0 1-.449-1.118c0-.449.15-.858.486-1.158.296-.296.672-.409 1.081-.409.446 0 .819.149 1.118.449.3.296.449.669.449 1.118.037.446-.113.819-.449 1.118Zm-2.013-2.05c-.262.26-.373.559-.373.895 0 .373.111.67.373.932.26.26.559.373.932.373s.672-.113.932-.373c.262-.262.373-.559.373-.932 0-.336-.111-.672-.373-.895-.26-.262-.559-.412-.932-.412-.373.039-.672.15-.932.412Zm1.418 1.714c-.15.113-.3.15-.449.15a.864.864 0 0 1-.633-.263.897.897 0 0 1-.262-.669c0-.263.076-.486.225-.672.147-.186.373-.26.633-.26.15 0 .299.037.449.11.186.113.333.3.373.599h-.412c-.037-.113-.074-.226-.15-.263-.074-.036-.147-.073-.223-.073a.34.34 0 0 0-.297.147.79.79 0 0 0-.113.449c0 .186.037.336.113.409.074.076.186.15.297.15.113 0 .186-.037.262-.074l.111-.225h.372a.726.726 0 0 1-.296.485Zm3.953-16.558c0-.149-.04-.298-.04-.447 0-.112-.036-.186-.036-.261a3.735 3.735 0 0 0-.15-.634 3.602 3.602 0 0 0-.669-1.156 3.742 3.742 0 0 0-.971-.783 5.049 5.049 0 0 0-1.043-.373c-.112 0-.189-.038-.262-.038.409-1.79-.113-3.17-.446-3.803l.073-.075c.446-.448.822-.858 1.155-1.268l.076-.074c.26-.373.41-.746.523-1.082.073-.298.149-.559.149-.821v-.708c-.039-.559-.113-1.268-.485-2.051a5.351 5.351 0 0 0-.783-1.156 4.607 4.607 0 0 0-1.231-.97c-.039-.037-.076-.037-.113-.074a18.281 18.281 0 0 0-3.131-3.469 22.982 22.982 0 0 0-1.418-1.118c-.409-.299-.745-.522-.932-.672a6.635 6.635 0 0 0-.932-.596 9.109 9.109 0 0 1-.784-.522c-.111-.112-.187-.15-.224-.187 0 0 .001-.037-.036-.112-.037-.112-.037-.261-.037-.522 0-.373 0-.932.037-1.678.073-1.193.186-2.797.336-4.997.036-.709.037-1.38.037-1.977a14.547 14.547 0 0 0-.373-2.722c-.037-.187-.113-.373-.15-.56-.037-.149-.076-.223-.113-.298a2.281 2.281 0 0 0-.149-.373c-.037-.074-.037-.149-.111-.224-.076-.186-.149-.373-.262-.559l-.037-.037-.037-.112-.036-.075a5.513 5.513 0 0 0-.636-.895 13.32 13.32 0 0 0-.632-2.125c-.299-.746-.672-1.567-1.232-2.387-.186-.299-.41-.56-.634-.821l.075-.074c.112-.112 2.723-2.872 4.81-6.452 3.843-6.675 2.575-10.74.821-12.978l-.037-.075-.036-.074c-1.68-1.865-3.993-2.983-6.789-3.356-5.855-.746-12.493 2.163-14.245 2.983l-3.208-3.506c-.298-.298-.596-.596-.932-.82a10.452 10.452 0 0 0-1.081-.634 8.545 8.545 0 0 0-2.462-.671c-1.753-.224-3.766 0-5.929.559-3.245.821-6.862 2.462-10.666 4.736-3.804 2.275-7.757 5.221-11.598 8.727-4.699 4.251-8.577 8.764-11.411 13.09-1.418 2.125-2.574 4.214-3.431 6.228-.858 1.976-1.418 3.878-1.641 5.594-.149 1.267-.149 2.461.112 3.542.111.485.26.933.484 1.38.261.56.597 1.082 1.007 1.529l3.319 3.468.373.373-.037.038-.112.223c-.074.112-.112.224-.186.299l-.112.223c-.038.112-.112.224-.15.336l-.111.224c-.038.112-.112.224-.149.336a.518.518 0 0 0-.075.223l-.112.336c-.037.074-.037.149-.075.224-.037.112-.074.224-.111.373 0 .037-.038.074-.038.111 0 .038-.037.075-.037.112-.037.112-.075.261-.112.373 0 .075-.037.149-.037.224 0 .037-.037.112-.037.149 0 .075-.038.149-.038.224 0 .074-.037.149-.037.224-.037.149-.037.298-.075.484 0 .038 0 .075-.037.15v.037c0 .112-.037.224-.037.298 0 .15-.038.336-.038.485v1.865c0 .074 0 .149.038.224 0 .074 0 .186.037.298v.112c0 .074.037.186.037.261 0 .074.038.149.038.261 0 .037 0 .074.037.112 0 .037.037.112.037.149v.074c0 .075.038.15.075.299 0 .037 0 .037.037.074.037.075.037.15.037.187l.038.075c0 .037 0 .037.037.074l.037.075v.112c0 .037 0 .037.038.074l.037.075c0 .074.037.111.037.186.037.075.037.149.075.298v-.037.037c0 .038 0 .038.037.075.037.112.075.187.075.224.037.074.074.149.112.298l.037.037c.037.075.037.15.074.187l.038.075c0 .037.037.074.037.111 0 .038.037.112.074.15l.038.074c.037.037.037.112.074.149.038.038.038.075.038.112.037.075.074.149.112.261l.037.037c.037.038.037.075.075.113.037.037.037.111.074.149a.598.598 0 0 1 .112.223l.037.038c.037.037.037.074.075.112.037.074.075.112.075.149.037.037.074.112.111.186v.037c.038.038.038.075.075.112a.507.507 0 0 0 .112.15c.037.037.074.111.149.223.037.038.075.075.075.112.037.075.074.112.111.149a.792.792 0 0 1 .112.187c.038.037.075.112.112.149a.293.293 0 0 0 .149.149l.112.112.038.037c.074.075.111.149.186.224l.074.075.112.111.224.224.075.075.112.112.037.037c.074.075.186.149.261.261l.037.037c.038.038.075.038.112.112.112.112.224.187.336.299l.074.037.038.037.074.038c.112.074.187.149.298.223l.038.037a.114.114 0 0 0 .075.038l.074.074c.075.075.149.112.224.187l.074.037.075.075.037.037c.075.037.187.112.261.149.038.037.075.075.149.075l.075.037.037.037h.038c.149.075.261.15.373.224l.074.037.448.224h.037l.074.038.224.111c.038 0 .075.038.112.038.037.037.075.037.149.074h.038a.518.518 0 0 0 .224.075h.037c.074.037.149.074.261.112h.037l.075.037c.037 0 .074.037.112.037h.037c.112.038.224.075.373.112h.037c.187.075.373.112.559.149h.038c.149.038.335.075.484.112h.15c.149.037.298.075.447.075h.149c.112 0 .187.037.299.037 0 .037.037.075.037.112.261.709.522 1.417.82 2.088 4.401 12.194 16.596 21.668 30.506 23.421 13.91 1.714 28.789-7.534 34.831-20.027.26-.485.486-.932.709-1.417.596-.634 1.081-1.268 1.381-1.939.483-.41.596-1.231.559-1.94Zm-77.234-4.848h.037-.037Z"
-              />
-              <path
-                fill="#132134"
-                d="M511.235 442.27a7.33 7.33 0 0 0-2.313-1.715 1.87 1.87 0 0 0-.486-.15c-.073-.037-.147-.111-.223-.149a9.887 9.887 0 0 0-2.349-.596c-4.028-.522-8.652.969-11.262 2.014-.15.074-.299.111-.448.186-.224.112-.447.186-.671.298-.373.15-.746.336-1.156.523-.075.037-.112.037-.187.074l-.708-.82-.336-.373-.298-.336-.112-.149-2.723-3.133-.223-.261-.299-.335h-.037l-.783.186-.895.187c-.485-.224-1.082-.373-1.753-.485-1.342-.187-3.058 0-4.997.485-1.939.484-4.14 1.342-6.452 2.498-4.661 2.312-9.808 5.855-14.805 10.405-1.753 1.566-3.356 3.207-4.848 4.811-1.454 1.566-2.76 3.17-3.953 4.698-2.424 3.17-4.252 6.228-5.445 8.988-.783 1.827-1.268 3.506-1.454 4.96-.149 1.044-.112 1.977.074 2.722.075.261.15.448.224.672.186.447.41.82.709 1.156l3.02 3.17.336.335.037-.037a1.13 1.13 0 0 1 .299-.298l.037-.038.112-.111c.074-.112.186-.187.261-.299.074-.074.112-.112.186-.149.075-.037.112-.112.187-.149l.074-.075c.075-.037.112-.112.187-.149.074-.075.186-.149.261-.224.037-.037.112-.074.149-.111h.037c.112-.075.187-.15.299-.224a.677.677 0 0 0 .186-.15c.112-.074.187-.149.298-.223.075-.037.112-.075.187-.112.112-.075.224-.149.336-.187.074-.037.111-.074.186-.111.112-.075.224-.112.336-.187.074-.037.112-.075.186-.112l.075-.037c.112-.037.186-.112.298-.149.037-.038.112-.038.149-.075.075-.037.15-.074.261-.112l.224-.112.075-.037.559-.224c.074-.037.112-.037.187-.074.074-.037.111-.037.186-.075l.075-.037c.037 0 .074-.037.111-.037.075-.038.15-.038.262-.075h.074c.112-.037.186-.037.299-.075.074-.037.186-.037.26-.074.112-.038.224-.038.336-.075.037 0 .075-.037.149-.037.038 0 .112 0 .149-.037.112-.038.224-.038.336-.038.112 0 .187-.037.298-.037.075 0 .15-.037.224-.037h.298c.373 0 .821 0 1.269.037h.074a12 12 0 0 1 1.492.224c.112.037.261.037.373.074.037-.037.037-.112.074-.149.299-.485.56-1.007.858-1.492a56.612 56.612 0 0 1 5.295-7.421c.635-.746 1.269-1.417 1.903-2.088a49.32 49.32 0 0 1 7.682-6.526c.522-.373 1.081-.709 1.566-1.045.709-.447 1.38-.82 2.014-1.193a19.573 19.573 0 0 1 1.492-.783c.783-.41 1.491-.746 2.125-1.007l.56-.224c.261-.112.522-.224.746-.298-.187-.037-.411-.037-.597-.037h-.746c-.186 0-.41 0-.597.037-.373.037-.783.074-1.193.112-.149.037-.261.037-.41.074 1.007-.522 2.014-.895 2.946-1.193a19.08 19.08 0 0 1 1.865-.448c.82-.149 1.603-.261 2.312-.298.485-.037.932-.037 1.342-.037.41 0 .821.037 1.156.037 1.231.075 1.977.261 1.977.261-.187-.261-.41-.447-.634-.634-.448-.336-.932-.559-1.38-.746-.186-.074-.336-.112-.485-.186-.112-.037-.224-.075-.298-.075.224-.037.41-.074.597-.074.223-.038.41-.038.634-.075 1.752-.149 3.319 0 4.475.187 1.38.26 2.237.596 2.237.596-.186-.373-.485-.708-.82-1.007a8.675 8.675 0 0 0-.709-.522c.485.112.895.224 1.343.373 1.119.336 2.051.746 2.797 1.082l.671.335c.074.038.112.075.149.075.672.373 1.045.634 1.045.634a2.82 2.82 0 0 0-.635-1.119c-.074-.074-.149-.187-.223-.261h.037c.037 0 .074.037.149.037.037 0 .112.037.187.075.298.112.745.298 1.268.559.037 0 .037.038.074.038l.224.111c.186.112.373.187.559.336l.56.336c.112.074.186.112.298.186.186.149.41.299.597.485.224.187.41.373.596.559l.112.112c.15.149.299.299.41.448.262.298.485.634.709.969.261.411.447.895.634 1.38 0 .038 0 .075.037.075.112-.112.336-.336.597-.634 2.164-2.536 8.653-10.741 4.737-15.738Z"
-              />
-              <path
-                fill="#90694C"
-                d="M520.892 503.542s.076-.559-.223-1.081c0-.038-.037-.075-.037-.075l-.037-.037-.113-.112a1.63 1.63 0 0 0-.26-.224c-.039-.037-.112-.074-.186-.112-.186-.074-.373-.149-.635-.186-.187-.037-.373-.037-.596-.037h-.299v.074c0 .038.036.075.036.149 0 .038 0 .075.04.075.037.112.073.298.11.447 0 0 0 .038-.037.112v.038c-.113.261-.409.857-1.081 1.603a3.743 3.743 0 0 1-.373.41c-1.864 1.94-5.893 4.662-14.395 6.192-2.872.409-5.557.446-7.794.149-3.357-.412-5.855-1.493-7.049-2.947-.112-.112-.186-.261-.298-.373-.075-.112-.112-.186-.187-.298-.037-.037-.037-.075-.074-.112a5.878 5.878 0 0 1-.709-2.797c0-1.156.373-2.237.97-3.058.485-.671 1.119-1.193 1.902-1.417.373-.112.783-.187 1.193-.149.075 0 .187 0 .336.037.261.037.671.074 1.156.149.037 0 .112 0 .149.038 1.604.186 4.14.447 7.198.26 5.593-.335 12.94-2.088 19.615-8.129v-.038c.037-.522-.036-1.417-.672-2.125a3.203 3.203 0 0 0-.372-.336c-.074-.074-.187-.112-.26-.186-.113-.038-.187-.112-.263-.15l-.333-.223v-.038c-.039-.074-.076-.149-.15-.223-1.716-2.611-4.103-4.14-5.071-4.848-.449-.336-.748-.485-.748-.485-2.05-1.268-2.35-1.678-2.163-5.072.076-1.231.186-2.834.336-5.035 0-.298.039-.559.039-.82v-1.081c0-.187-.039-.373-.039-.523 0-.074 0-.149-.037-.223-.037-.187-.037-.373-.074-.56 0-.074-.039-.186-.039-.261 0-.112-.037-.186-.073-.298-.037-.075-.037-.187-.074-.261a1.047 1.047 0 0 1-.076-.336c0-.037-.037-.074-.037-.112 0-.037-.037-.074-.037-.112v-.037l-.036-.112a.495.495 0 0 1-.076-.224v-.037l-.037-.074-.037-.075c-.076-.149-.113-.298-.186-.41l-.037-.112c-.076-.112-.15-.261-.226-.373l-.037-.037-.036-.038c-.076-.074-.113-.149-.187-.223l-.075-.112a1.59 1.59 0 0 0-.223-.261v-.299a10.59 10.59 0 0 1-.112-.522c0-.037-.038-.111-.038-.149v-.037c0-.075-.037-.112-.037-.187-.037-.149-.074-.298-.149-.447 0-.037-.037-.112-.037-.149l-.112-.336c-.038-.075-.038-.149-.075-.186a1.302 1.302 0 0 0-.149-.336l-.112-.224c-.037-.112-.112-.223-.149-.335l-.112-.224c-.075-.112-.112-.224-.186-.336l-.112-.224a3.008 3.008 0 0 1-.224-.373c-.037-.074-.075-.111-.112-.186-.037-.037-.074-.112-.112-.149-.112-.112-.186-.224-.298-.336l-.075-.037c-.149-.149-.261-.298-.41-.41l-.037-.038c-.075-.037-.149-.111-.187-.149h-.037c-.074-.037-.112-.074-.186-.112l-.038-.037c-.074-.037-.186-.112-.335-.186-.038-.038-.112-.038-.187-.075-.335-.112-.746-.261-1.268-.41-.037 0-.074 0-.112-.037-.41-.112-.895-.224-1.417-.299-.186-.037-.335-.037-.522-.074-.075 0-.149 0-.261-.038h-.112c-.41-.037-.783-.074-1.193-.074h-.037c-.522 0-1.045.037-1.529.149.111-.186.447-.41.745-.597.411-.224.784-.373.784-.373-.224 0-.448.038-.672.075-.074 0-.149.037-.223.037-.038 0-.112 0-.15.037h-.037c-.037 0-.074 0-.112.038a19.269 19.269 0 0 0-2.909.932l-.223.112h-.038c-.074 0-.111 0-.186.037a3.305 3.305 0 0 1 1.641-1.529c.074-.037.186-.074.261-.112.037 0 .037 0 .074-.037.075-.037.149-.037.187-.075.112-.037.186-.037.186-.037h-.596a8.901 8.901 0 0 0-3.394 1.119c-.671.373-1.268.783-1.79 1.193-.149.038-.261.075-.41.075-1.417.298-3.021.671-4.699 1.193-.634.187-1.268.41-1.902.634-.485.187-1.007.373-1.492.56-.373.149-.709.298-1.081.447-2.574 1.156-5.147 2.685-7.272 4.736l-.224.224-.56.56-.223.223c-.075.112-.187.224-.261.299-.411.447-.783.932-1.156 1.454a12.74 12.74 0 0 0-.746 1.193l-.038.038c-.484.857-.932 1.753-1.268 2.759-.186.485-.335 1.007-.485 1.529-.037.149-.074.261-.111.411l-.038.074c-.149.56-.261 1.156-.335 1.753 0 .112-.037.224-.037.373a7.137 7.137 0 0 0-.075.932c-.037.298-.037.56-.037.858v1.156c0 .336 0 .671.037 1.007v.037c0 .336.037.671.075 1.044 0 .149.037.299.037.448v.112c0 .074 0 .149.037.261.037.373.112.783.149 1.193 0 .112.038.224.038.336.074.335.111.708.186 1.044 0 0-.075-.448-.037-1.119 0-.149.037-.335.037-.485.037-.372.112-.82.261-1.23l.112-.336c0-.037.037-.075.074-.149.075-.186.187-.41.299-.597.037.149.037.299.075.411v1.976c.037 1.268.149 2.797.41 4.102 0 0 .037-2.349.447-2.648 0 0 .709 4.401 1.343 6.042.074.223.186.373.261.485v-.448l-.075-2.424v-.261a43.146 43.146 0 0 0 1.678 3.095c.187.299.336.597.522.895-1.193 1.044-2.312 2.574-3.132 4.364-.112.298-.261.596-.373.895-.075.186-.112.335-.187.522 0 .037-.037.112-.037.149-.037.112-.074.261-.112.373 0 .074-.037.112-.037.186-.037.112-.075.261-.112.373 0 .075-.037.112-.037.187l-.112.559a2.896 2.896 0 0 0-.074.559c0 .075 0 .112-.038.187 0 .149-.037.261-.037.41 0 .075 0 .149-.038.224 0 .149-.037.261-.037.41v1.305c0 .15 0 .299.037.485v.112c0 .186.038.373.075.597 0 .074 0 .112.038.186 0 .149.037.261.074.41 0 .075.037.15.037.224.038.112.038.261.075.373 0 .075.037.149.037.224.037.112.075.224.075.373.037.074.037.149.074.261.038.112.075.261.112.373.037.076.037.149.075.223.037.112.074.262.112.372.037.076.037.15.074.226.038.15.112.297.149.41.038.073.038.112.075.186l.559 1.231c.038.074.075.11.112.186l.224.447a.79.79 0 0 0 .149.225c.075.111.149.26.224.373.037.074.112.147.149.223.075.111.149.26.261.373.075.074.112.15.187.26.074.113.186.262.261.373.074.076.111.149.186.262.112.113.187.26.298.373.075.073.112.15.187.223.112.15.224.263.373.41l.186.189c.187.186.373.409.56.595 0 0-.112-.149-.261-.448a.28.28 0 0 1-.075-.187l-.112-.223a1.934 1.934 0 0 1-.112-.26 6.268 6.268 0 0 1-.522-2.685l.336.446c.112.263.298.599.485.971.074.111.149.26.186.373.299.522.597 1.118.97 1.677.447.709.895 1.381 1.38 2.014 0 0-.56-1.192-.746-1.977-.038-.223-.075-.409-.038-.522.038.037.075.113.15.149.037.037.074.113.149.15.074.113.186.26.335.446l.075.076c.075.074.112.187.187.26.074.113.149.226.26.336.784 1.008 1.865 2.35 2.611 3.058.261.224.485.41.634.486l-1.082-1.94c2.462 2.014 12.68 4.586 22.712 1.717 1.231-.336 2.424-.785 3.617-1.307.709-.297 1.417-.67 2.125-1.042l.113-.077c.485-.259.968-.559 1.454-.858.373-.223.746-.483 1.082-.709a27.329 27.329 0 0 0 1.79-1.341c.037-.037.073-.074.113-.074a21.24 21.24 0 0 0 1.604-1.53 22.548 22.548 0 0 0 1.787-2.126l.04-.037c.147-.186.259-.336.372-.522l.037-.037c.037-.074.074-.113.113-.186.186-.26.333-.486.52-.746.039-.076.076-.113.113-.186.073-.15.149-.262.262-.411.074-.112.11-.187.186-.298.074-.112.147-.224.187-.336.11-.186.223-.41.333-.597l.113-.224.223-.447.113-.224.037-.074c.076-.187.186-.373.262-.56v-.037c.334-.299.559-.597.783-.858.036-.037.073-.074.073-.112.226-.708.373-1.268.373-1.79Zm-60.003-7.235c0-.037 0-.074-.038-.149-.112-.634-.223-1.305-.261-1.976v-.149c-.112.037-.223.186-.372.373-.038.074-.113.149-.15.261-.335.634-.634 1.529-.634 1.529.038-.299.075-.597.075-.895v-.112c.037-.299.074-.634.112-.933v-.112c.037-.373.112-.708.186-1.044 0-.037 0-.074.038-.149v-.037c0-.037 0-.075.037-.112 0-.075.037-.112.037-.187.037-.261.112-.522.149-.783.037-.224.075-.447.112-.634v-.074c0-.037 0-.075.038-.075v-.074c0-.038.037-.112.037-.15 0-.037 0-.074.037-.111 0-.075.037-.149.037-.187 0-.037.038-.112.038-.149 0-.075.037-.149.037-.187 0-.037 0-.074.037-.111 0-.038 0-.075.037-.075 0-.075.038-.112.038-.149v-.112l-.075.074c-.074.038-.112.112-.186.15l-.075.074a.99.99 0 0 0-.186.224l-.038.037-.223.336-.038.037c0 .038-.037.038-.037.075l-.112.224v.074c-.037.112-.112.224-.149.336-.112.298-.224.596-.298.858-.038.074-.038.186-.075.223-.037.149-.075.299-.075.336v.037-.037c0-.112 0-.187.038-.298 0-.187.037-.411.037-.597 0-.149.037-.336.075-.485 0-.037 0-.112.037-.149 0-.075.037-.149.037-.224 0-.074.038-.149.038-.224v-.111c0-.075.037-.187.037-.262v-.261c0-.037 0-.074.037-.111.037-.112.075-.262.075-.373v-.038c.037-.074.037-.186.074-.261v-.037c.038-.186.112-.373.149-.559v-.075c0-.037 0-.075.038-.112.074-.298.186-.597.261-.858v-.037c.074-.224.186-.485.261-.708v-.038c.037-.112.075-.186.112-.298.074-.149.112-.298.186-.447 0-.038.037-.112.112-.262.037-.037.037-.111.075-.149a1.79 1.79 0 0 0 .111-.261l.038-.075.634-1.156c.037-.037.037-.111.074-.149v-.037l.038-.074c.074-.112.149-.262.223-.373.112-.187.187-.374.299-.523.186-.335.335-.634.447-.783a.408.408 0 0 1 .112-.186c-.037 0-.224-.149-.522-.336a17.043 17.043 0 0 0-1.529-.82c-.224-.112-.41-.187-.634-.299a21.924 21.924 0 0 0-.858-.335c-.074-.038-.149-.038-.223-.075-.261-.074-.485-.149-.746-.223-.187-.038-.373-.075-.522-.113-.15-.037-.299-.037-.448-.074-.074 0-.149 0-.224-.037-.074 0-.149 0-.223-.038-.187 0-.411-.037-.597-.037h-.299a6.93 6.93 0 0 0-1.267.112h-.075c-.746.112-1.454.298-2.163.559-.224.075-.448.187-.708.299-.075.037-.112.074-.187.074l-.224.112c-.671.336-1.305.783-1.902 1.268h-.037c-.522.41-.97.895-1.38 1.38l-.336.448c-.149.186-.26.372-.372.559-.224.335-.448.708-.634 1.081-.075.149-.15.261-.187.411 0 .037-.037.037-.037.074-.373.895-.671 1.827-.783 2.834-.037.187-.037.373-.075.597v.075c0 .149-.037.298-.037.484v.672c0 .149 0 .298.037.447v.187c0 .111.038.224.038.335v.038c0 .074 0 .111.037.186 0 .112.037.224.037.299 0 .074.037.149.037.223.038.112.038.187.075.299a.53.53 0 0 0 .074.223v.038l.038.074v.075l.037.074c.037.075.037.15.075.224.037.075.037.186.074.261a.532.532 0 0 1 .075.261l.037.075c.037.074.037.112.075.186.037.075.074.149.112.261l.111.224c.038.037.038.075.075.112.037.037.037.075.075.112l.112.224c.037.074.074.186.149.261.037.074.074.149.112.186.037.075.111.186.149.261.037.075.074.112.112.187.037.074.112.186.186.261.037.074.075.111.112.149.074.074.112.186.187.261.037.037.074.112.111.149.075.075.149.187.224.261.037.037.075.075.112.149.075.075.149.187.224.261l.112.112c.074.075.149.187.261.261l.112.112.261.261c.037.037.074.075.111.075.075.074.187.149.299.261.037.037.074.037.112.074.111.075.186.15.298.224.037.037.075.037.075.075l.335.223.075.038.335.224.075.037c.112.074.261.149.373.223l.075.038c.111.075.261.149.373.186l.074.038c.149.074.261.111.41.186l.038.037c.149.075.298.112.41.187h.037c.149.037.298.112.448.149h.037c.075.037.149.037.224.075.074.037.149.037.261.074h.037c.149.037.336.075.485.112.149.037.335.075.485.075a22.4 22.4 0 0 0 1.566.149h.373c.186 0 .335 0 .522-.037.224-.038.485-.038.746-.075.223-.037.447-.075.634-.112.335-.075.671-.149.969-.261.97-.261 1.79-.634 2.313-1.044 0 0-.336-1.492-.597-2.983-.075-.075-.112-.224-.149-.374Z"
-              />
-              <path
-                fill="#2F83C1"
-                d="m484.831 440.89-.112-1.491.746-1.641.298-.075c-.484-.223-1.081-.373-1.753-.484-1.342-.187-3.057 0-4.997.484-1.939.485-4.139 1.343-6.451 2.499-4.662 2.312-9.808 5.855-14.805 10.405-1.753 1.566-3.357 3.207-4.849 4.811-1.454 1.566-2.759 3.169-3.953 4.698-2.424 3.17-4.251 6.228-5.444 8.988-.784 1.827-1.269 3.505-1.455 4.96-.149 1.044-.112 1.976.075 2.722.074.261.149.448.224.672.186.447.41.82.708 1.156l3.021 3.17.335.335.038-.037c.074-.112.186-.224.298-.298l.037-.038.112-.112c.075-.112.187-.186.261-.298.075-.075.112-.112.187-.149.074-.038.112-.112.186-.149l.075-.075c.074-.037.112-.112.186-.149.075-.075.187-.149.261-.224.038-.037.112-.075.149-.112h.038c.111-.074.186-.149.298-.224a.628.628 0 0 0 .186-.149c.112-.074.187-.149.299-.224.075-.037.112-.074.186-.111.112-.075.224-.149.336-.187.075-.037.112-.074.186-.112.112-.074.224-.112.336-.186.075-.037.112-.075.187-.112l.074-.037c.709-1.977 1.79-4.177 3.207-6.415 1.716-2.759 3.916-5.668 6.527-8.577a72.884 72.884 0 0 1 5.37-5.407 79.202 79.202 0 0 1 7.607-6.116c1.716-1.194 3.394-2.276 5.035-3.208.149-.074.298-.186.485-.261.447-.261.858-.485 1.305-.708a35.986 35.986 0 0 1 2.76-1.305l-.448-.523-.074-.074-.56-.634-3.095-3.729.037-2.686-.037-.037 2.2-.671 4.177 2.125Z"
-              />
-              <path
-                fill="#FCE9CD"
-                d="M520.894 503.543s.074-.56-.225-1.082c0-.037-.037-.075-.037-.075l-.037-.037-.113-.112a1.546 1.546 0 0 0-.26-.224c-.036-.037-.113-.074-.186-.111a2.562 2.562 0 0 0-.635-.187c-.187-.037-.373-.037-.596-.037h-.299v.075c0 .037.039.074.039.149 0 .037 0 .074.037.074.036.112.073.299.11.448 0 0 0 .037-.037.112v.037c-.11.261-.409.858-1.081 1.603-.111.15-.223.262-.373.411-1.864 1.939-5.893 4.661-14.395 6.191-2.871.409-5.556.446-7.794.149-3.356-.412-5.855-1.493-7.048-2.947-.112-.111-.187-.261-.299-.373-.074-.112-.112-.186-.186-.298-.037-.037-.037-.074-.075-.112a5.889 5.889 0 0 1-.708-2.797c0-1.156.373-2.237.969-3.058.485-.671 1.119-1.193 1.902-1.417.373-.112.783-.186 1.194-.149.074 0 .186 0 .335.037.261.037.672.075 1.156.15.038 0 .112 0 .15.037 1.603.186 4.139.447 7.197.261 5.594-.336 12.94-2.089 19.615-8.13v-.037c.04-.523-.036-1.417-.669-2.126a3.69 3.69 0 0 0-.373-.336c-.076-.074-.189-.111-.262-.186-.113-.038-.187-.112-.26-.149l-.336-.224v-.037a.778.778 0 0 0-.15-.224c-1.717-2.611-4.103-4.14-5.071-4.848-.449-.336-.746-.485-.746-.485-2.053-1.268-2.349-1.678-2.163-5.072.074-1.231.187-2.834.336-5.034 0-.299.037-.56.037-.821v-1.081c0-.187-.037-.373-.037-.522 0-.075 0-.15-.039-.224-.037-.187-.037-.373-.074-.56 0-.074-.036-.186-.036-.261 0-.111-.037-.186-.077-.298-.036-.075-.036-.186-.073-.261-.037-.112-.076-.224-.076-.336 0-.037-.037-.074-.037-.112 0-.037-.037-.074-.037-.111v-.038l-.036-.111c-.037-.075-.077-.15-.077-.224v-.038l-.036-.074-.037-.075c-.074-.149-.113-.298-.186-.41l-.037-.112c-.076-.112-.15-.261-.223-.373l-.04-.037-.036-.037c-.074-.075-.111-.15-.187-.224l-.074-.112a1.592 1.592 0 0 0-.224-.261l-.298-.299c-.112-.074-.224-.149-.336-.186-.037-.037-.074-.037-.112-.037l-.223-.112c-.038 0-.075-.037-.112-.037-.075-.038-.15-.038-.187-.075-.037 0-.074-.037-.149-.037-.074 0-.112-.038-.187-.038h-.82c-.037 0-.074 0-.112.038-.037 0-.074 0-.112.037-.074 0-.149.037-.261.075h-.074c-.224.074-.411.149-.634.223l-.075.038c-.075.037-.149.074-.261.111l-.112.038-.223.112-.113.074-.223.112-.112.075c-.075.037-.149.112-.224.149l-.074.037c-.075.075-.187.112-.261.187l-.038.037c-.112.075-.186.149-.298.186-.186.112-.373.262-.56.373a5.48 5.48 0 0 1-.484.336l-.075.037a3.44 3.44 0 0 1-.485.299l-.074.037c-.149.112-.336.187-.485.298l-.075.038c-.149.111-.335.186-.522.261l-.074.037c-.187.075-.336.149-.522.224l-.075.037c-.187.075-.373.112-.559.186h-.038c-.186.038-.41.075-.596.112a9.66 9.66 0 0 1-1.045.075c-.559 0-1.118-.037-1.715-.112l-.559-.112c-.075 0-.15-.037-.187-.037-.485-.112-1.007-.224-1.492-.336-.223-.037-.41-.112-.634-.149a47.642 47.642 0 0 0-1.044-.224c-.074 0-.149-.037-.186-.037-.038 0-.112-.037-.149-.037-.075 0-.15-.037-.224-.037-.075 0-.149-.038-.224-.038h-.075c-.074 0-.149-.037-.186-.037a1.855 1.855 0 0 0-.485-.075c-.037 0-.112 0-.149-.037-.112 0-.224-.037-.336-.037-.074 0-.112 0-.186-.038-.112 0-.187-.037-.298-.037h-1.716c-.187 0-.336.037-.522.037h-.149c-.112 0-.224.038-.336.038h-.037c-.075 0-.112.037-.187.037-.112.037-.223.037-.335.074-.075 0-.149.038-.187.038-.112.037-.223.075-.335.075-.075 0-.112.037-.187.074-.112.037-.261.075-.373.112-.074.037-.112.037-.149.074-.186.075-.373.15-.522.224a11.85 11.85 0 0 0-3.58 3.207c-.336.448-.634.895-.895 1.343-.298.485-.522 1.007-.746 1.529-.075.149-.112.298-.186.447-.075.262-.187.523-.262.784 0 .037 0 .074-.037.074-.074.298-.149.597-.224.933a9.01 9.01 0 0 0-.149.969c-.037.187-.037.336-.037.522a9.058 9.058 0 0 0 0 1.417v.112c0 .112 0 .224.037.373.038.336.075.671.149 1.007.038.112.075.224.075.373.037.112.037.224.075.336a8.61 8.61 0 0 0 .969 2.312c1.529 2.648 4.065 4.251 5.93 5.109.708.298 1.454.597 2.237.858l-.074.074a8.857 8.857 0 0 1-.485.112c-.261.075-.597.187-.932.299-.373.111-.821.298-1.268.484-1.194.485-2.536 1.194-3.73 2.163-.149.112-.298.261-.447.373a6.414 6.414 0 0 0-.709.709l-.074.074-.112.112c-.075.112-.187.187-.261.299l-.112.112c-.112.149-.224.298-.298.447-.112.149-.187.298-.299.447-.708 1.231-1.118 2.723-1.007 4.513v-.075c0-.037 0-.074.038-.112v-.037c0-.037.037-.074.037-.149v-.037c0-.038.037-.112.037-.187 0-.074.037-.112.075-.186v-.038c.037-.074.037-.149.074-.223v-.037c.038-.075.038-.15.075-.224v-.038c.037-.074.075-.149.112-.261l.037-.037c.037-.075.075-.187.149-.261.075-.149.149-.261.224-.41.037-.038.074-.112.112-.149l.037-.038.224-.335c.075-.075.149-.187.224-.261l.074-.075c.075-.074.112-.149.187-.187l.074-.074c.075-.075.187-.149.262-.261 0 0-1.604 3.654-.896 6.526V503.319c0-.075 0-.15.038-.224 0-.075 0-.149.037-.224v-.037c0-.075.037-.187.037-.261v-.075c.038-.112.038-.186.075-.298v-.037c.037-.112.074-.224.074-.299 0-.037.038-.074.038-.111l.037-.075c.037-.075.075-.187.112-.261.037-.112.112-.224.186-.336 0-.037.038-.075.038-.075l.224-.335s-.075.895-.075 1.976c0 .187 0 .336.037.522 0 .113 0 .224.038.336-.038.746-.038 1.492.037 2.275.112 1.305.447 2.722 1.007 4.176 0 0-.523-1.677.223-3.355l.224.447c.075.485.149 1.194.299 1.977.149 1.304.745 5.22 3.43 7.943 0 0-1.007-1.305-.857-2.982.112.223.223.409.298.595.448.859 1.343 2.423 2.349 3.544 0 0-.783-1.53-.596-1.866 0 0 2.61 3.244 3.431 3.544l-1.082-1.791-.037-.036c3.766 2.758 8.988 3.766 15.402 1.976 1.119-.299 2.238-.709 3.431-1.194.708-.299 1.417-.672 2.126-1.045l.112-.073c.483-.263.969-.56 1.455-.859.372-.223.745-.485 1.081-.709a28.741 28.741 0 0 0 1.79-1.341c.037-.037.074-.076.111-.076a20.639 20.639 0 0 0 1.604-1.528 22.595 22.595 0 0 0 1.79-2.126l.037-.037c.149-.186.262-.336.372-.522l.04-.037c.036-.076.073-.113.11-.186.186-.263.336-.486.522-.746.037-.076.074-.113.113-.186.074-.151.147-.262.26-.412.076-.111.113-.186.186-.298.077-.112.15-.224.187-.335.113-.187.223-.411.336-.597l.113-.224.223-.448.113-.223.036-.075c.074-.186.187-.373.26-.559v-.037c.336-.299.559-.597.785-.858.037-.038.074-.075.074-.112a2.76 2.76 0 0 0 .485-1.566Zm-60.303-9.51c-.112.037-.224.186-.373.373-.038.074-.112.149-.149.261-.336.634-.634 1.529-.634 1.529.037-.299.074-.597.074-.895v-.112c.038-.298.075-.634.112-.932v-.112c.037-.373.112-.709.186-1.044 0-.038 0-.075.038-.15v-.037c0-.037 0-.074.037-.112 0-.075.038-.112.038-.186.037-.261.111-.522.149-.783.037-.224.074-.448.111-.634v-.075c0-.037 0-.075.038-.075v-.074c0-.037.037-.112.037-.149 0-.038 0-.075.037-.112 0-.075.038-.15.038-.187 0-.037.037-.112.037-.149 0-.074.037-.149.037-.186 0-.038 0-.075.038-.112 0-.037 0-.075.037-.075 0-.074.037-.112.037-.149v-.112l-.074.075c-.075.037-.112.112-.187.149l-.075.075a.99.99 0 0 0-.186.223l-.037.038-.224.335-.037.038c0 .037-.038.037-.038.074l-.111.224v.074c-.038.112-.112.224-.15.336-.112.299-.223.597-.298.858-.037.074-.037.186-.075.224-.037.149-.074.298-.074.335v.038-.038c0-.112 0-.186.037-.298 0-.187.037-.41.037-.597 0-.149.038-.336.075-.485 0-.037 0-.111.037-.149 0-.074.038-.149.038-.224 0-.074.037-.149.037-.223v-.112c0-.075.037-.187.037-.261v-.261c0-.038 0-.075.037-.112.038-.112.075-.261.075-.373v-.037c.037-.075.037-.187.075-.262v-.037c.037-.186.112-.373.149-.559v-.075c0-.037 0-.074.037-.112.075-.298.186-.596.261-.857v-.038c.075-.223.187-.485.261-.708v-.037c.037-.112.075-.187.112-.299.075-.149.112-.298.187-.447 0-.038.037-.112.111-.261.038-.038.038-.112.075-.15.037-.074.075-.149.112-.261l.037-.074c-.037-.038-.112-.038-.149-.075h-.037l-.075-.037a9.233 9.233 0 0 0-1.081-.41h-.075a5.183 5.183 0 0 0-.746-.149c-.149 0-.298-.038-.447-.038h-.373c-.187 0-.373.038-.56.038-.074 0-.112 0-.186.037-.075 0-.187.037-.261.037-.037 0-.075 0-.112.037-.074 0-.112.038-.187.038-.111.037-.261.074-.372.111-.038 0-.112.038-.15.038-.149.075-.335.112-.485.186-.074.038-.111.075-.186.075l-.224.112-.074.037a.786.786 0 0 0-.224.149c-.075.038-.149.075-.224.149-.074.038-.149.112-.224.15-.037.037-.111.074-.149.111-.112.112-.261.224-.373.336l-.186.187c-.112.111-.186.223-.298.335l-.038.038c-.037.074-.112.111-.149.186l-.112.112a6.555 6.555 0 0 0-.522.858c-.037.037-.037.074-.074.111-.075.15-.112.299-.187.448-.037.075-.037.112-.075.186-.037.15-.111.299-.149.485v.15c-.037.111-.037.223-.074.335 0 .075 0 .112-.037.187a.34.34 0 0 1-.038.186v.149c0 .186-.037.373 0 .56 0 .149 0 .298.038.447v.075c0 .074.037.149.037.224.037.298.112.559.186.857.075.224.15.41.224.597.149.373.336.708.559 1.044a5.927 5.927 0 0 0 1.753 1.715c.373.262.82.448 1.268.597.41.149.858.261 1.343.299.112 0 .223.037.298.037h.485c.149 0 .298-.037.447-.037.075 0 .187 0 .261-.038.224-.037.411-.074.634-.112h.038c.074-.037.149-.037.261-.074.261-.075.522-.187.746-.299l.037-.037a.628.628 0 0 1 .186-.074c-.112-.635-.223-1.306-.261-1.977.187-.075.187-.112.187-.149Z"
-              />
-              <path fill="#4D9ED7" d="m485.763 441.338.112.074-.112-.074Z" />
-              <path
-                fill="#fff"
-                d="M518.619 502.013c0-.037 0-.074-.037-.074-.039-.075-.039-.112-.039-.149a5.374 5.374 0 0 0-.633-1.38 3.932 3.932 0 0 0-.486-.634l.113-.448c.037-.186.074-.373.074-.559.149-1.343-.187-2.387-.483-2.984a2.318 2.318 0 0 0-.113-.261c-.223.15-.41.299-.635.448-4.027 2.722-8.539 4.475-13.424 5.221-3.357.522-6.9.559-10.592.112l.075.037s-2.238.298-2.834 2.424l.037.037-.187 1.417c.075 1.94.933 3.133 1.94 3.767 1.044.41 2.312.709 3.729.896 2.424.296 5.258.223 8.092-.223 3.804-.524 9.174-2.538 13.017-5.148.855-.597 1.677-1.231 2.349-1.865l.037-.037.036-.037c0-.261-.036-.41-.036-.56Z"
-              />
-              <path
-                fill="#F2D2AC"
-                d="M511.347 483.591c-2.053-1.268-2.349-1.679-2.163-5.072.074-1.231.187-2.834.334-5.035 0-.298.039-.559.039-.82v-1.082c0-.186-.039-.373-.039-.522 0-.074 0-.149-.037-.224-.037-.186-.037-.373-.074-.559 0-.075-.036-.186-.036-.261 0-.112-.04-.187-.077-.298-.036-.075-.036-.187-.073-.262a5.192 5.192 0 0 0-1.045-1.156l-.074-.074-.075-.037h-.037c-.075-.038-.112-.075-.187-.112-.074-.038-.149-.075-.186-.112-.597-.373-1.305-.597-2.014-.709-.41-.074-.82-.074-1.231-.074-.186 0-.41.037-.596.037-.523.074-1.045.187-1.604.336-.186.111-.373.261-.559.373-.149.111-.299.223-.485.335l-.075.037a3.44 3.44 0 0 1-.485.299l-.074.037c-.149.112-.336.187-.485.298l-.075.038c-.149.112-.335.186-.522.261l-.074.037c-.187.075-.336.149-.522.224l-.075.037c-.187.075-.373.112-.559.187h-.038c-.186.037-.41.074-.596.111a9.66 9.66 0 0 1-1.044.075c-.56 0-1.119-.037-1.716-.112l-.559-.112c-.075 0-.15-.037-.187-.037-.485-.112-1.007-.224-1.491-.336-.224-.037-.411-.111-.635-.149-.335-.074-.671-.149-1.044-.224-.074 0-.149-.037-.186-.037-.298 0-.597.037-.858.075h-.075c-.857.112-1.678.298-2.535.597-.038 0-.075.037-.149.037-.15.037-.262.112-.411.149-.186.074-.335.149-.522.224-1.23.522-2.387 1.268-3.506 2.2-.447.373-.857.783-1.267 1.193a16.55 16.55 0 0 0-1.604 2.014c-.037.037-.037.075-.075.112a9.313 9.313 0 0 0-.671 1.156 4.1 4.1 0 0 1-.223.41v.038c-.224.484-.411.932-.56 1.417 0 .037 0 .074-.037.074-.038.075-.038.149-.075.224-.074.224-.149.448-.186.671a3.91 3.91 0 0 0-.149.709c0 .037-.038.112-.038.149-.037.149-.037.298-.074.41-.075.522-.112 1.045-.15 1.567v.372c.224.858.56 1.641.97 2.313 1.529 2.647 4.065 4.251 5.93 5.109.708.298 1.454.596 2.237.858l.038.037.037-.037c.112-.075.186-.15.298-.224.038-.038.075-.038.112-.075l.336-.223c.037-.038.074-.038.112-.075l.335-.224c.037-.037.075-.037.112-.074.112-.075.224-.15.373-.224.037-.038.075-.038.112-.075.112-.075.261-.149.373-.224.037-.037.074-.037.112-.074.111-.075.261-.149.373-.224.037-.037.111-.037.149-.075.112-.074.261-.149.373-.223a.278.278 0 0 1 .149-.075c.112-.074.261-.149.41-.224.075-.037.112-.074.187-.074.149-.075.261-.149.41-.187.074-.037.112-.074.186-.074.149-.075.261-.112.41-.187l.224-.112c.149-.074.261-.111.41-.186.075-.037.149-.075.262-.112.149-.037.261-.112.41-.149.074-.037.186-.075.261-.112.149-.037.261-.112.41-.149.112-.037.186-.075.298-.112.149-.037.261-.112.411-.149l.335-.112c.112-.037.261-.075.373-.149.112-.038.261-.075.373-.112.112-.037.261-.075.373-.112.149-.037.298-.075.41-.112.112-.037.224-.075.373-.112.149-.037.298-.074.485-.149.074-.037.149-.037.224-.074.596-1.53 2.424-5.333 5.519-5.744.336-.037.634-.074.932-.074h.224c-.037-.224-.075-.448-.075-.709-.111-1.641.411-3.021 1.119-3.058.56-.037 1.007.783 1.082 1.044.224.485.373 1.119.41 1.828a6.554 6.554 0 0 1-.149 1.753c.074.074.112.149.186.186.709.895.821 2.088.783 2.983.261-.037.56-.037.858-.074.074 0 .15 0 .223-.038h.15a9.942 9.942 0 0 1 1.081-.074 52.13 52.13 0 0 1 2.499-.149c-.485-.261-.782-.41-.782-.41Zm7.794 22.972a23.218 23.218 0 0 1-4.177 2.946 33.9 33.9 0 0 1-2.089 1.006 27.307 27.307 0 0 1-4.289 1.454c-2.871.709-4.848.709-4.848.709 1.567.858 3.543.858 4.884.748.935-.076 1.568-.263 1.568-.263-.037.56-.111 1.082-.187 1.604a7.66 7.66 0 0 1-.446 1.491v.037c-.226.562-.522 1.045-.822 1.494-.036.037-.073.11-.113.149l.113-.076c.486-.26.969-.559 1.455-.858.372-.223.745-.483 1.081-.709a27.42 27.42 0 0 0 1.79-1.341c.037-.037.074-.074.113-.074a21.24 21.24 0 0 0 1.604-1.53 23.034 23.034 0 0 0 1.79-2.126l.037-.037c.15-.187.26-.336.373-.522l.037-.037c.036-.074.073-.113.112-.187.187-.26.334-.485.523-.745.037-.076.073-.113.11-.187.076-.149.15-.261.263-.41.073-.112.11-.187.186-.299.073-.111.147-.223.186-.335.111-.187.223-.41.336-.597l.111-.224.223-.447.113-.224.036-.074c.076-.187.187-.373.263-.56-.15.038-.226.112-.336.224Zm-34.012-15.029c-.373.112-.82.298-1.268.485-1.193.484-2.536 1.193-3.729 2.163-.149.112-.298.261-.448.373a6.311 6.311 0 0 0-.708.708l-.075.075-.111.112c-.075.112-.187.186-.262.298l-.111.112c-.112.149-.224.298-.299.447-.112.149-.186.299-.298.448-.709 1.23-1.119 2.722-1.007 4.512v-.074c0-.038 0-.075.037-.112v-.037c0-.038.037-.075.037-.149v-.038c0-.037.038-.112.038-.186 0-.075.037-.112.074-.187v-.037c.038-.075.038-.149.075-.224v-.037c.037-.075.037-.149.075-.224v-.037c.037-.075.074-.149.111-.261l.038-.038c.037-.074.074-.186.149-.261.074-.149.149-.261.224-.41.037-.037.074-.112.112-.149l.037-.037.224-.336c.074-.075.149-.186.223-.261l.075-.075c.075-.074.112-.149.186-.186l.075-.075c.074-.074.187-.149.261-.261 0 0-1.604 3.655-.895 6.526V503.542c0-.074 0-.149.037-.224 0-.074 0-.149.038-.223v-.037c0-.075.037-.187.037-.262v-.074c.037-.112.037-.187.074-.299v-.037c.038-.112.075-.224.075-.298 0-.037.037-.075.037-.112l.038-.075c.037-.074.074-.186.112-.261.037-.112.111-.223.186-.335 0-.038.037-.075.037-.075l.224-.335s-.074.894-.074 1.976c0 .187 0 .336.037.522 0 .112 0 .224.037.336.075-1.156.298-2.238.597-3.245.074-.261.186-.559.261-.82 0-.038.037-.075.037-.112.112-.261.187-.485.298-.746.038-.037.038-.074.075-.149.112-.224.224-.485.336-.709.037-.037.037-.074.037-.112.261-.484.522-.932.783-1.342.037-.038.037-.075.075-.112.111-.149.223-.298.335-.485.038-.037.038-.075.075-.112.037-.037.037-.074.074-.112.15-.186.262-.373.411-.559.037-.037.037-.075.074-.075.149-.186.298-.373.448-.522.186-.223.373-.41.559-.634.149-.149.261-.298.41-.41l.15-.149c.298-.298.596-.597.894-.858a16.457 16.457 0 0 1 1.716-1.38 8.646 8.646 0 0 1-.485.112 5.758 5.758 0 0 1-.895.336Z"
-              />
-              <path
-                fill="#F2D2AC"
-                d="M506.61 483.442s0-.075-.037-.224c-.149-.522-.671-1.79-2.723-1.604-2.163.224-2.908 1.977-3.095 2.574-.037.111-.037.186-.037.186l.074.037c.261-.223.784-.223 1.268.112.224.149.373.298.522.485.187.298.224.597.112.82l-.112.112s.075.038.224.075c.597.186 2.275.522 3.17-.634.037-.075.075-.112.112-.187-.186-.074-.261-.335-.261-.671 0-.149.037-.298.075-.485.111-.447.447-.708.708-.596Zm-10.778 15.961a76.377 76.377 0 0 1-3.58.671c.038 0 .112 0 .149.038 1.604.186 4.14.447 7.198.26 5.594-.335 12.94-2.088 19.615-8.129v-.038c.037-.522-.036-1.417-.672-2.125a3.203 3.203 0 0 0-.372-.336c-.037 0-.037-.037-.074-.037-.186-.112-.336-.187-.485-.299l-.336-.223v-.038c-.41.597-.896 1.157-1.415 1.679-.187.186-.373.373-.599.559-1.827 1.641-4.213 3.058-6.785 4.177-4.253 1.864-9.026 3.095-12.644 3.841Zm-35.353-9.733c-.075.037-.112.111-.187.149l-.074.074c-.075.075-.149.15-.187.224l-.037.037-.224.336-.037.037c0 .037-.037.037-.037.075l-.112.224v.074c-.038.112-.112.224-.149.336-.112.298-.224.597-.299.858-.037.074-.037.186-.074.223a2.96 2.96 0 0 0-.075.336v.037-.037c0-.112 0-.186.037-.298 0-.187.038-.411.038-.597a2.2 2.2 0 0 1 .074-.485c0-.037 0-.112.038-.149 0-.075.037-.149.037-.224 0-.074.037-.149.037-.223v-.113c0-.074.037-.186.037-.26v-.261c0-.038 0-.075.038-.113.037-.111.074-.26.074-.372v-.038c.038-.074.038-.186.075-.261v-.037c.037-.186.112-.373.149-.559v-.075c0-.037 0-.074.038-.112.074-.298.186-.596.26-.858v-.037c.075-.224.187-.485.262-.708v-.038c.037-.112.074-.186.111-.298.075-.149.112-.298.187-.448 0-.037.037-.111.112-.261.037-.037.037-.111.074-.149.037-.075.075-.149.112-.261l.038-.074c-.038-.038-.113-.038-.15-.075h-.037l-.075-.037a9.101 9.101 0 0 0-1.081-.411h-.075a5.487 5.487 0 0 0-.745-.149c-.15 0-.299-.037-.448-.037h-.373c-.186 0-.373.037-.559.037-.075 0-.112 0-.187.038-.074 0-.186.037-.261.037-.037 0-.075 0-.112.037-.074 0-.112.037-.186.037-.112.038-.261.075-.373.112-.037 0-.112.038-.149.038-.149.074-.336.111-.485.186-.075.037-.112.075-.187.075l-.223.112-.075.037a.766.766 0 0 0-.224.149c-.074.037-.149.074-.224.149-.074.038-.149.112-.223.149-.037.038-.112.075-.149.112-.112.112-.261.224-.373.336l-.187.186c-.112.112-.186.224-.298.336l-.037.037c-.038.075-.112.112-.15.187l-.111.112a6.501 6.501 0 0 0-.522.857c-.038.038-.038.075-.075.112-.075.149-.112.299-.187.448-.037.074-.037.112-.074.186 0 .037.037.075.037.112.037.149.075.261.112.41v.075c.037.112.112.261.149.373.149.373.298.708.485 1.007.037.037.037.074.075.112v.037l.037.037c.112.224.224.41.335.56.187.261.336.447.448.596l.075.075c0 .037.037.037.037.037l.074.075c0-.075-.037-.149-.037-.187a6.096 6.096 0 0 1-.149-1.417v-.075c0-.298 0-.596.037-.857l.112.336c.037.149.112.298.149.447v.074a7.69 7.69 0 0 0 .746 1.567c.299.522.56.858.746 1.081.037.038.075.075.075.112l.074.075c-.112-.522-.186-1.007-.186-1.492v-.485c0-.223 0-.41.037-.596.075.335.187.671.298.969v.038a10.911 10.911 0 0 0 1.231 2.349l.037.037a.766.766 0 0 1 .149.224c.038.037.075.075.075.112l.075.075.037.037v-.075c0-.074-.037-.149-.037-.224-.038-.261-.075-.522-.112-.745 0-.15-.038-.336-.038-.485v-.821c0-.186.038-.41.038-.596.037.112.074.261.112.373.037.074.074.186.112.261v.037l.037.037c.037.112.074.224.112.298.074.187.149.373.261.523.112.223.186.41.298.559.075.112.112.187.149.261.075.112.15.224.187.298a.8.8 0 0 1 .149.224l.037.038c.037.074.112.111.149.186.038.037.075.074.075.112.075.075.112.149.149.149l.037.037v-.111c.038-.299.075-.635.112-.933v-.112c.038-.373.112-.708.187-1.044 0-.037 0-.074.037-.149v-.038c0-.037 0-.074.038-.111 0-.075.037-.112.037-.187.037-.261.112-.522.149-.783.037-.224.074-.447.112-.634v-.075c0-.037 0-.074.037-.074v-.075c0-.037.037-.111.037-.149 0-.037 0-.075.038-.112 0-.074.037-.149.037-.186 0-.037.037-.112.037-.149 0-.075.038-.149.038-.187 0-.037 0-.074.037-.112 0-.037 0-.074.037-.074 0-.075.037-.112.037-.149v-.112l.187.112Zm45.199 15.588c.522-.187.969-.336 1.342-.485.709-.298 1.082-.522 1.082-.597-.037-.074-.448-.037-1.194.15-.373.074-.82.149-1.379.298-.261.074-.56.112-.858.149-.299.038-.634.112-.97.149-.671.112-1.417.187-2.237.261l-1.231.112c-.41.038-.858.038-1.305.075-1.79.037-3.767-.037-5.781-.41-1.044-.187-2.498-.672-3.505-1.231l-.038-.037c.597-2.126 2.835-2.424 2.835-2.424l-.075-.038h-.112c-.335-.037-.596-.074-.858-.111-.186-.038-.335-.038-.447-.075-.112 0-.224-.037-.261-.037-1.342-.075-2.088 1.044-2.312 1.528l-.112.224a4.599 4.599 0 0 0-.373 1.492 4.143 4.143 0 0 0 .97 3.021c.522.634 1.417 1.193 2.573 1.641-1.007-.634-1.902-1.865-1.939-3.767 1.156.485 2.461.783 3.468.97 2.126.41 4.214.447 6.041.335.448-.037.933-.074 1.343-.112.447-.074.858-.111 1.268-.149.82-.149 1.529-.261 2.237-.41.336-.075.672-.149.97-.261.336-.074.597-.149.858-.261Z"
-              />
-              <path
-                fill="#664F44"
-                d="M503.925 480.346c1.119.038 1.864.373 2.387.895a6.31 6.31 0 0 0 .149-1.753c-.038-.708-.187-1.342-.41-1.827-.075-.298-.523-1.081-1.082-1.044-.709.037-1.231 1.417-1.119 3.058 0 .224.038.447.075.671Zm1.864 4.215c0 .298.112.559.262.671.261.149.596-.149.745-.634 0-.037.038-.075.038-.112.111-.448 0-.895-.262-1.007-.26-.149-.596.149-.745.634 0 .149-.038.298-.038.448Zm-5.071.82c.111.186.298.373.522.522.485.298 1.007.336 1.305.074l.112-.111c.112-.224.074-.522-.112-.821a1.311 1.311 0 0 0-.522-.484c-.485-.299-1.007-.336-1.268-.113-.037.038-.075.075-.112.15-.112.186-.074.484.075.783Zm18.533 6.824c-.037 0-.037 0 0 0Zm-.596 10.33a17.74 17.74 0 0 1-2.386 1.903c-3.84 2.61-9.212 4.661-13.016 5.145-2.871.449-5.668.523-8.092.223-1.417-.186-2.648-.483-3.729-.893-1.156-.448-2.051-1.007-2.574-1.641-.783-.933-1.007-2.051-.969-3.021.037-.522.149-1.044.373-1.492l.112-.224c.261-.484.969-1.64 2.312-1.529.037 0 .149 0 .261.038.112 0 .298.037.447.074.261.038.523.075.858.112h.112c3.692.448 7.235.411 10.591-.112 4.886-.746 9.398-2.498 13.424-5.221.226-.149.413-.298.636-.447l2.199-3.282c-6.675 6.079-14.021 7.832-19.615 8.167a36.57 36.57 0 0 1-7.198-.261c-.074 0-.112 0-.149-.037-.485-.075-.858-.112-1.156-.149-.149 0-.261-.038-.335-.038-.411-.037-.821.038-1.194.15-.746.223-1.417.745-1.902 1.417-.597.82-.97 1.939-.97 3.058 0 .932.224 1.902.709 2.797.037.037.037.074.075.112.074.111.112.186.186.298.075.112.186.261.298.373 1.157 1.455 3.692 2.5 7.049 2.946 2.275.299 4.922.262 7.794-.15 8.502-1.528 12.531-4.288 14.395-6.19.15-.149.26-.261.373-.41.672-.783.968-1.38 1.081-1.604v-.037c.037-.075.037-.112.037-.112l-.037.037Z"
-              />
-              <path
-                fill="#E9B784"
-                d="M491.619 473.074a3.412 3.412 0 0 0-.858-.336 20.373 20.373 0 0 0-.597-.112c-.074 0-.149 0-.224-.037-1.006-.075-1.864.075-1.864.075-.187.037-.373.074-.522.149-1.343.41-2.275 1.118-2.946 1.939-.522.671-.895 1.38-1.082 2.088-.037.075-.037.112-.037.187-.037.186-.112.373-.149.522-.112.597-.187 1.119-.187 1.566 0 .597.037.97.037.97.224-.522.485-.97.746-1.417.634-1.082 1.306-1.977 2.014-2.685 1.343-1.38 2.723-2.126 3.767-2.499 1.119-.373 1.902-.41 1.902-.41Zm16.148-1.902-.038-.075a2.527 2.527 0 0 0-.559-.522c-.522-.336-1.231-.522-2.2-.336 0 0-1.156.262-1.828 1.045 0 0 2.014-.858 4.177.671.186.149.41.298.597.485.186.186.41.373.597.634 0 0-.074-.933-.635-1.753-.037-.075-.074-.112-.111-.149Z"
-              />
-              <path
-                fill="#664F44"
-                d="M491.879 486.015a60.086 60.086 0 0 1 2.537-.672c.149-.708.186-1.566.074-2.498-.261-2.163-1.231-3.841-2.2-3.729-.97.111-1.529 1.976-1.268 4.139.149 1.119.447 2.088.857 2.76Z"
-              />
-              <path
-                fill="#FFD250"
-                d="m487.441 437.273-2.013.448.41 3.654-5.333-2.722-2.2.671.037.037 6.489 7.384.074-.037.075-.037c.187-.112.41-.261.597-.373.074-.038.149-.075.224-.149.298-.187.596-.373.895-.523l-2.76-3.169 4.438 2.163c.149-.075.261-.15.41-.224l-.41-3.17 2.013 2.349h.038c.186-.074.335-.186.522-.261l.074-.037.075-.037c.037 0 .075-.037.075-.037l.223-.113s.038 0 .038-.037c.224-.112.485-.223.708-.335l-4.699-5.445Z"
-              />
-              <path
-                fill="#4D9ED7"
-                d="M485.465 437.832v-.112.112Zm0-.112.298-.074c-.485-.224-1.081-.373-1.753-.485-1.342-.186-3.057 0-4.997.485-1.939.485-4.139 1.342-6.451 2.499-4.662 2.312-9.808 5.855-14.806 10.404-1.753 1.567-3.356 3.207-4.848 4.811l6.042 7.458c1.641-1.789 3.393-3.579 5.332-5.332 2.574-2.35 5.184-4.438 7.758-6.228a61.03 61.03 0 0 1 7.346-4.438l-1.044-4.774.037-2.685-.037-.037 2.2-.671 4.289 2.2-.112-1.492.746-1.641Z"
-              />
-              <path
-                fill="#132134"
-                d="m485.465 437.72-.746 1.641.112 1.529 1.044.522-.41-3.692Zm-1.529 4.736 4.475 2.201-1.715.969-2.76-3.17Zm4.475-1.23.41 3.17 1.604-.821-2.014-2.349Z"
-              />
-              <path
-                fill="#fff"
-                d="M505.267 440.07h.821c.745 0 1.454.112 2.125.261.037 0 .074 0 .113.037h.037l.073.038h.076c-.076-.038-.149-.112-.225-.149a9.83 9.83 0 0 0-2.349-.597c-4.027-.522-8.651.97-11.262 2.014-.149.074-.299.112-.448.186-.223.112-.447.187-.671.298-.373.15-.746.336-1.156.522-.074.038-.112.038-.187.075-.223.112-.484.224-.708.336 0 0-.037 0-.037.037l-.224.112c-.037 0-.074.037-.074.037h-.15l-.037.038a3.345 3.345 0 0 0-.522.261h-.038c-.037.037-.074.037-.111.074h-.038c-.037 0-.074.037-.074.037-.261.15-.522.262-.783.411-.187.112-.373.186-.56.298h-.037c-.149.075-.261.149-.41.224-.448.223-.858.485-1.306.746-.037.037-.074.037-.149.074-.037.038-.074.038-.111.075-.038.037-.075.037-.113.074l-.037.038c-.298.186-.596.335-.895.522-.074.037-.149.074-.224.149-.186.112-.41.261-.596.373l-.075.037-.074.038-1.679 1.118a72.157 72.157 0 0 0-2.722 1.939c-.112.075-.224.187-.336.262-2.349 1.752-4.587 3.654-6.75 5.593-2.274 2.089-4.475 4.215-6.526 6.378a87.89 87.89 0 0 0-2.834 2.946c-1.566 1.678-2.424 2.722-2.387 2.759 0 0 .149-.111.41-.373-.186.224-.41.485-.596.709 3.431-3.655 7.197-7.384 11.262-10.889a109.738 109.738 0 0 1 7.683-6.079 117.518 117.518 0 0 1 3.542-2.424c.373-.261.784-.485 1.156-.746.187-.112.373-.224.597-.336.112-.074.224-.149.336-.186.224-.112.41-.261.634-.373.186-.112.41-.224.596-.373.112-.075.224-.112.336-.187l.783-.447c.149-.075.299-.187.448-.261a53.5 53.5 0 0 1 1.491-.783c.038-.037.075-.037.112-.075l.224-.112c.112-.037.187-.112.298-.149.038-.037.075-.037.112-.074l.224-.112s.037 0 .037-.038h.038c.223-.112.447-.223.634-.298.037 0 .074-.037.111-.075l.224-.111c.261-.112.485-.224.746-.336a.795.795 0 0 0 .186-.112c.038 0 .075-.037.112-.037a.527.527 0 0 1 .224-.075c.298-.112.56-.261.858-.373.261-.112.485-.223.746-.336a54.813 54.813 0 0 1 3.953-1.491c2.312-.634 4.363-1.044 6.153-1.119Z"
-              />
-            </g>
-            <g filter="url(#filter5_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="247.8"
-                y="484.05"
-                fill="#0E1521"
-                rx="64"
-              />
-              <path
-                fill="#25F4EE"
-                d="M305.91 541.249v-3.478a25.718 25.718 0 0 0-3.655-.314c-11.851-.025-22.344 7.643-25.903 18.929-3.558 11.287.642 23.573 10.371 30.331a26.993 26.993 0 0 1-5.167-28.943 27.068 27.068 0 0 1 24.354-16.525Z"
-              />
-              <path
-                fill="#25F4EE"
-                d="M306.567 580.645c6.626-.009 12.073-5.221 12.363-11.83v-58.981h10.792a20.237 20.237 0 0 1-.314-3.734h-14.761v58.923c-.246 6.643-5.705 11.905-12.363 11.916a12.581 12.581 0 0 1-5.711-1.425 12.404 12.404 0 0 0 9.994 5.131ZM349.88 529.846v-3.279a20.174 20.174 0 0 1-11.164-3.363 20.485 20.485 0 0 0 11.164 6.642Z"
-              />
-              <path
-                fill="#FE2C55"
-                d="M338.716 523.204a20.278 20.278 0 0 1-5.053-13.398h-3.941a20.413 20.413 0 0 0 8.994 13.398ZM302.255 552.195a12.386 12.386 0 0 0-11.973 9.433 12.363 12.363 0 0 0 6.263 13.885 12.327 12.327 0 0 1-.997-12.856 12.366 12.366 0 0 1 11.019-6.728c1.239.016 2.469.208 3.654.57v-14.994a25.991 25.991 0 0 0-3.654-.285h-.657v11.402a12.764 12.764 0 0 0-3.655-.427Z"
-              />
-              <path
-                fill="#FE2C55"
-                d="M349.88 529.846v11.402a35.072 35.072 0 0 1-20.472-6.642v29.961c-.031 14.95-12.179 27.053-27.153 27.053a26.892 26.892 0 0 1-15.532-4.903 27.125 27.125 0 0 0 29.753 6.757c10.367-4.066 17.184-14.052 17.187-25.173v-29.875a35.124 35.124 0 0 0 20.5 6.557v-14.681a21.044 21.044 0 0 1-4.283-.456Z"
-              />
-              <path
-                fill="#000"
-                d="M329.408 564.567v-29.96a35.072 35.072 0 0 0 20.501 6.556V529.76a20.478 20.478 0 0 1-11.193-6.556 20.413 20.413 0 0 1-8.994-13.398H318.93v59.009a12.377 12.377 0 0 1-8.8 11.338 12.404 12.404 0 0 1-13.614-4.582 12.366 12.366 0 0 1-6.263-13.885 12.388 12.388 0 0 1 11.974-9.434c1.239.011 2.47.203 3.654.57V541.42c-10.731.181-20.341 6.676-24.497 16.555a26.994 26.994 0 0 0 5.31 29.056 26.901 26.901 0 0 0 15.561 4.589c14.974 0 27.122-12.103 27.153-27.053Z"
-              />
-            </g>
-            <g filter="url(#filter6_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="413.7"
-                y="82.95"
-                fill="#0E1521"
-                rx="64"
-              />
-              <path
-                fill="#1080F8"
-                d="M478.423 131.283c-7.188-9.153-13.199-12.641-20.393-12.641-14.667 0-25.905 19.088-25.905 39.291 0 12.643 6.116 20.616 16.361 20.616 7.373 0 12.676-3.475 22.103-19.955 0 0 3.93-6.939 6.634-11.72a307.406 307.406 0 0 1 2.993 4.944l4.421 7.437c8.611 14.411 13.41 19.294 22.103 19.294 9.981 0 15.535-8.082 15.535-20.988 0-21.153-11.491-38.919-25.451-38.919-7.392 0-13.17 5.568-18.401 12.641Zm13.825 21.341-5.286-8.815a178.39 178.39 0 0 0-4.124-6.424c4.764-7.353 8.694-11.017 13.367-11.017 9.709 0 17.477 14.295 17.477 31.854 0 6.693-2.193 10.577-6.735 10.577-4.353 0-6.433-2.875-14.699-16.175Zm-50.373 5.474c0-14.667 7.313-29.664 16.031-29.664 4.72 0 8.665 2.726 14.708 11.377-5.738 8.801-9.214 14.321-9.214 14.321-7.643 11.981-10.287 14.667-14.542 14.667-4.38 0-6.983-3.845-6.983-10.701Z"
-              />
-              <path
-                fill="url(#paint0_linear_11024_3769)"
-                d="M432.125 157.933c0-20.203 11.238-39.291 25.905-39.291l-.124 9.792c-8.718 0-16.031 14.997-16.031 29.664 0 6.856 2.603 10.701 6.983 10.701l-.372 9.75c-10.245 0-16.361-7.973-16.361-20.616Z"
-              />
-              <path
-                fill="url(#paint1_linear_11024_3769)"
-                d="M458.03 118.642c7.194 0 13.205 3.488 20.393 12.641 3.122 3.975 5.894 8.223 8.539 12.526l5.286 8.815c8.266 13.3 10.346 16.175 14.699 16.175l-.207 9.75c-8.693 0-13.492-4.883-22.103-19.294l-4.421-7.437a307.406 307.406 0 0 0-2.993-4.944c-.906-1.462-3.013-4.779-4.609-7.063-6.043-8.651-9.988-11.377-14.708-11.377l.124-9.792Z"
-              />
-              <path
-                fill="url(#paint2_linear_11024_3769)"
-                d="M522.275 157.562c0 12.905-5.554 20.987-15.535 20.987l.207-9.75c4.542 0 6.735-3.884 6.735-10.577l8.593-.66Z"
-              />
-            </g>
-            <g filter="url(#filter7_dd_11024_3769)">
-              <rect
-                width="134.4"
-                height="134.4"
-                x="12.6"
-                y="248.85"
-                fill="#0E1521"
-                rx="64"
-              />
-              <path
-                fill="#1877F2"
-                fillRule="evenodd"
-                d="M107.446 287.7c-9.767 0-17.406 2.811-23.446 7.268 5.95 5.081 10.404 10.593 13.454 22.132 8.412-1.556 25.455-5.725 32.746-29.4h-22.754ZM90.404 315.934c-5.521-15.86-19.19-27.361-37.382-28.165-.308-.044-24.672-.069-24.672-.069 8.175 27.888 36.068 29.98 36.068 29.98v44.57H97.65s-1.77-30.587-7.246-46.316Z"
-                clipRule="evenodd"
-              />
-            </g>
-            <g filter="url(#filter8_dd_11024_3769)">
-              <rect
-                width="168"
-                height="168"
-                x="231"
-                y="232.05"
-                fill="#0E1521"
-                rx="80"
-              />
-              <g fill="#1FEC90" clipPath="url(#clip1_11024_3769)">
-                <path d="M310.505 303.44c3.193 8.02 3.284 17.094.092 25.113l24.977-.075v-25.03l-25.069-.008ZM310.505 303.44l17.721-17.691-17.74-17.699-15.775 15.739 8.253 8.234a33.773 33.773 0 0 1 7.541 11.417Z" />
-                <path d="m302.964 339.903-8.253 8.234 15.775 15.739 17.74-17.699-17.629-17.624c-1.656 4.154-4.263 7.988-7.633 11.35Z" />
-              </g>
-            </g>
-          </g>
-          <defs>
-            <filter
-              id="filter0_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="49.9"
-              y="50.95"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter1_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="451"
-              y="216.85"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter2_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="49.9"
-              y="382.75"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter3_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="215.8"
-              y="-18.35"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter4_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="381.7"
-              y="382.75"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter5_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="215.8"
-              y="452.05"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter6_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="381.7"
-              y="50.95"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter7_dd_11024_3769"
-              width="198.4"
-              height="198.4"
-              x="-19.4"
-              y="216.85"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="4" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <filter
-              id="filter8_dd_11024_3769"
-              width="232"
-              height="232"
-              x="199"
-              y="200.05"
-              color-interpolation-filters="sRGB"
-              filterUnits="userSpaceOnUse"
-            >
-              <feFlood flood-opacity="0" result="BackgroundImageFix" />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset dy="4" />
-              <feGaussianBlur stdDeviation="6" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0" />
-              <feBlend
-                in2="BackgroundImageFix"
-                result="effect1_dropShadow_11024_3769"
-              />
-              <feColorMatrix
-                in="SourceAlpha"
-                result="hardAlpha"
-                values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"
-              />
-              <feOffset />
-              <feGaussianBlur stdDeviation="16" />
-              <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.05 0" />
-              <feBlend
-                in2="effect1_dropShadow_11024_3769"
-                result="effect2_dropShadow_11024_3769"
-              />
-              <feBlend
-                in="SourceGraphic"
-                in2="effect2_dropShadow_11024_3769"
-                result="shape"
-              />
-            </filter>
-            <linearGradient
-              id="paint0_linear_11024_3769"
-              x1="458.03"
-              x2="455.487"
-              y1="128.44"
-              y2="157.497"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stop-color="#0C64DD" />
-              <stop offset="1" stop-color="#1080F8" />
-            </linearGradient>
-            <linearGradient
-              id="paint1_linear_11024_3769"
-              x1="478.423"
-              x2="506.947"
-              y1="131.413"
-              y2="168.786"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stop-color="#0C64DD" />
-              <stop offset="1" stop-color="#1076EE" />
-            </linearGradient>
-            <linearGradient
-              id="paint2_linear_11024_3769"
-              x1="506.939"
-              x2="513.96"
-              y1="168.795"
-              y2="159.001"
-              gradientUnits="userSpaceOnUse"
-            >
-              <stop stop-color="#1077EF" />
-              <stop offset="1" stop-color="#1080F8" stop-opacity="0" />
-            </linearGradient>
-            <clipPath id="clip0_11024_3769">
-              <path fill="#fff" d="M0 0h630v630H0z" />
-            </clipPath>
-            <clipPath id="clip1_11024_3769">
-              <path fill="#fff" d="M267 268.05h96v96h-96z" />
-            </clipPath>
-          </defs>
-        </svg>
-      </div>
-      <div className="flex flex-col justify-center pl-24">
-        <p className="font-mono text-sm m-0 p-0 text-left">
-          ELASTIC PATH COMPOSER
-        </p>
-        <h2 className="mb-4 text-[2.5rem] text-white font-semibold tracking-tight">
-          Quickly connect commerce data to every tool in your arsenal
-        </h2>
-        <p className="text-sm m-0 p-0 text-left">
-          Composer comes preconfigured with dozens of instant-on integrations.
-          Need more? Quickly customize how data flows between
-        </p>
-        <div className="flex text-black gap-8 mt-8">
-          <button className="bg-transparent tracking-tight rounded-full ring-2 ring-gray-300 py-2 px-4 hover:ring-green-500 transition-all border-0 text-xs font-semibold">
-            30+ Instant-on integrations
-          </button>
-          <button className="bg-transparent tracking-tight rounded-full ring-2 ring-gray-300 py-2 px-4 hover:ring-green-500 transition-all border-0 text-xs font-semibold">
-            Or build your own
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
